Add test + fix

+++ vm/src/vm/instruction/execution.rs
@@ -117,7 +117,7 @@ impl ArithOp {
             ArithOp::SetLessThan => (a < b) as i32,
             ArithOp::SetLessThanU => ((a as u32) < (b as u32)) as i32,
             ArithOp::Mul => (a as i64 * b as i64) as i32,
-            ArithOp::MulHigh => ((a as i64 * b as i64) >> 32) as i32,
+            ArithOp::MulHigh => ((((a as u32) as u64) * ((b as u32) as u64)) >> 32) as i32,
             ArithOp::MulHighSignedUnsigned => ((a as i64 * (b as u32) as i64) >> 32) as i32,
             ArithOp::MulHighUnsigned => ((a as u64 * b as u64) >> 32) as i32,
             ArithOp::Div => {
